Move shared types from flowrlib to flowcore (#1707)
Move Block, Job/Payload, DebugCommand/BreakpointSpec, services, and
discovery modules from flowrlib to flowcore. Re-exports in flowrlib
maintain backward compatibility for existing code.

- Block and Job are always available (runtime needs them)
- DebugCommand is behind the debugger feature flag
- discovery is gated behind cfg(not(wasm32)) since mdns-sd doesn't
  support WASM targets
- services module always available, DEBUG_SERVICE_NAME behind debugger

flowdblib now imports Block, Job, DebugCommand from flowcore instead
of flowrlib. Still depends on flowrlib for RunState, State, and
DebuggerHandler which remain there.
