+++ flowcore/Cargo.toml
@@ -42,6 +42,7 @@ serde_yml = "0.0.13"
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 curl = {version = "~0.4" }
 simpath = { version = "~2.5", features = ["urls"] }
+mdns-sd = { version = "0.20", default-features = false }
 
 [lints]
 workspace = true
