+++ flowdb/src/lib/debug_handler.rs
@@ -5,13 +5,13 @@
 
 use serde_json::Value;
 
+use flowcore::model::block::Block;
+use flowcore::model::debug_command::DebugCommand;
 use flowcore::model::input::Input;
+use flowcore::model::job::Job;
 use flowcore::model::output_connection::OutputConnection;
 use flowcore::model::runtime_function::RuntimeFunction;
-use flowrlib::block::Block;
-use flowrlib::debug_command::DebugCommand;
 use flowrlib::debugger_handler::DebuggerHandler;
-use flowrlib::job::Job;
 use flowrlib::run_state::{RunState, State};
 
 use crate::coordinator_connection::{CoordinatorConnection, WAIT};
